server/requirements.txt
anaconda==0.0.1.1
asgiref==3.4.1
awscli==1.20.43
boto3==1.18.43
botocore==1.21.43
centroid-tracker==0.0.9
certifi==2021.5.30
charset-normalizer==2.0.5
click==8.0.1
colorama==0.4.3
Django==3.2.7
django-cors-headers==3.8.0
djangorestframework==3.12.4
dlib==19.22.1
docker==3.4.0
docker-pycreds==0.4.0
docutils==0.15.2
greenlet==1.1.1
idna==3.2
imutils==0.5.4
Jinja2==3.0.1
jmespath==0.10.0
keyring==8.7
keyrings.alt==4.1.0
MarkupSafe==2.0.1
numpy==1.21.2
opencv-ython4.5.3.56
pandas==1.3.3
psycopg2-binary==2.9.1
pyasn1==0.4.8
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
rsa==4.7.2
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
scipy==1.7.1
six==1.16.0
SQLAlchemy==1.4.23
sqlparse==0.4.2
tabulate==0.8.9
urllib3==1.26.6
voluptuous==0.12.1
websocket-client==1.2.1
yolo==0.3.2
passlib==1.7.2
bcrypt==3.2.0